ia-conversational/requirements.txt
# IA Conversational Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.12

# LangChain and AI
langchain==0.0.335
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langchain-community==0.0.6
openai==1.3.7
anthropic==0.7.8

# RAG and Embeddings
chromadb==0.4.18
qdrant-client==1.6.9
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Intent Classification and NLP
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
spacy==3.7.2
textblob==0.17.1

# Sentiment Analysis
vaderSentiment==3.3.2

# Database
supabase==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.3
numpy==1.25.2

# Async support
asyncio==3.4.3

# Environment and utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0

# Development and testing dependencies
isort==5.12.0
pytest-cov==4.1.0
factory-boy==3.3.0
